Update the migrations to pass SQLite env
See #6

+++ database/migrations/2016_09_15_103057_add_seller_buyer_and_details_columns_to_transactions_table.php
@@ -14,16 +14,18 @@ class AddSellerBuyerAndDetailsColumnsToTransactionsTable extends Migration
     public function up()
     {
         Schema::table('transactions', function (Blueprint $table) {
-            $table->dropForeign('transactions_user_id_foreign');
-            $table->dropColumn(['user_id', 'data']);
+            if(!env('DB_CONNECTION') == 'sqlite') {
+                $table->dropForeign('transactions_user_id_foreign');
+                $table->dropColumn(['user_id', 'data']);
 
-            $table->unsignedInteger('buyer_id')->after('ad_id');
-            $table->unsignedInteger('seller_id')->after('buyer_id');
-            $table->json('stripe_data')->nullable()->after('seller_id');
-            $table->json('details')->nullable()->after('stripe_data');
+                $table->unsignedInteger('buyer_id')->nullable()->after('ad_id');
+                $table->unsignedInteger('seller_id')->nullable()->after('buyer_id');
+                $table->json('stripe_data')->nullable()->after('seller_id');
+                $table->json('details')->nullable()->after('stripe_data');
 
-            $table->foreign('buyer_id')->references('id')->on('users');
-            $table->foreign('seller_id')->references('id')->on('users');
+                $table->foreign('buyer_id')->references('id')->on('users');
+                $table->foreign('seller_id')->references('id')->on('users');
+            }
         });
     }
 
@@ -35,13 +37,15 @@ class AddSellerBuyerAndDetailsColumnsToTransactionsTable extends Migration
     public function down()
     {
         Schema::table('transactions', function (Blueprint $table) {
-            $table->dropForeign('transactions_buyer_id_foreign');
-            $table->dropForeign('transactions_seller_id_foreign');
-            $table->dropColumn(['buyer_id', 'seller_id', 'details']);
+            if(!env('DB_CONNECTION') == 'sqlite') {
+                $table->dropForeign('transactions_buyer_id_foreign');
+                $table->dropForeign('transactions_seller_id_foreign');
+                $table->dropColumn(['buyer_id', 'seller_id', 'details']);
 
-            $table->unsignedInteger('user_id')->after('id');
-            $table->text('data')->after('user_id');
-            $table->foreign('user_id')->references('id')->on('users');
+                $table->unsignedInteger('user_id')->after('id');
+                $table->text('data')->after('user_id');
+                $table->foreign('user_id')->references('id')->on('users');
+            }
         });
     }
 }
